Moved source files to own src folder.
