+++ package.json
@@ -2,8 +2,9 @@
   "name": "dank-times-bot",
   "version": "1.1.0",
   "description": "This Telegram bot keeps track of 'dank' times such as 13:37 and 04:20, and awards points to users who call them out.",
-  "main": "main.js",
+  "main": "src/main.js",
   "scripts": {
+    "start": "node src/main.js",
     "test": "mocha"
   },
   "author": "Agadar",
